Remove Old Dev UI: gRPC
