+++ extensions/grpc/deployment/pom.xml
@@ -57,13 +57,7 @@
             <artifactId>quarkus-grpc-codegen</artifactId>
         </dependency>
 
-        <!-- for dev mode -->
         <dependency>
-            <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-vertx-http-dev-console-spi</artifactId>
-        </dependency>
-
-       <dependency>
             <groupId>io.quarkus</groupId>
             <artifactId>quarkus-resteasy-reactive-deployment</artifactId>
             <scope>test</scope>
