+++ extensions/grpc/deployment/src/main/java/io/quarkus/grpc/deployment/FieldDefinalizingVisitor.java
@@ -1,4 +1,4 @@
-package io.quarkus.grpc.deployment.devmode;
+package io.quarkus.grpc.deployment;
 
 import static java.util.Arrays.asList;
