--- extensions/grpc/deployment/src/main/resources/dev-templates/embedded.html
@@ -1,3 +1,0 @@
-<a href="{urlbase}/services" class="badge badge-light">
-  <i class="fas fa-cogs"></i>
-  Services</a>
